requirements.txt
black==23.10.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==2.3.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0
ua-parser==0.18.0
urllib3==2.0.7
user-agents==2.2.0
Werkzeug==2.3.0
zipp==3.17.0
